Move `getExtendedType` from `effectiveness.ts` to `type.ts`
